llm_model/requirements.txt
langchain
tavily-python
python-dotenv
requests
